Default storage is now 'file' rather than FFS This is exactly what OpenStack needs.
Therefore remove the openstack storage as well
